+++ packages/plugin-hardhat/test/transparent-admin-initial-owner.js
@@ -3,10 +3,14 @@ const test = require('ava');
 const { ethers, upgrades } = require('hardhat');
 const hre = require('hardhat');
 
+const ProxyAdmin = require('@openzeppelin/upgrades-core/artifacts/@openzeppelin/contracts-v5/proxy/transparent/ProxyAdmin.sol/ProxyAdmin.json');
+
 const OWNABLE_ABI = ['function owner() view returns (address)'];
 
 test.before(async t => {
   t.context.Greeter = await ethers.getContractFactory('Greeter');
+  t.context.HasOwner = await ethers.getContractFactory('HasOwner');
+  t.context.ProxyAdmin = await ethers.getContractFactory(ProxyAdmin.abi, ProxyAdmin.bytecode);
 });
 
 test('initial owner using default signer', async t => {
@@ -61,4 +65,45 @@ test('initial owner - no signer in ContractFactory', async t => {
   const admin = await hre.ethers.getContractAt(OWNABLE_ABI, adminAddress);
 
   t.is(await admin.owner(), initialOwner.address);
+});
+
+test('initial owner - must not be ProxyAdmin ', async t => {
+  const { Greeter, ProxyAdmin } = t.context;
+
+  const defaultSigner = await ethers.provider.getSigner(0);
+  const proxyAdmin = await ProxyAdmin.deploy(defaultSigner.address);
+  const predeployedProxyAdminAddress = await proxyAdmin.getAddress();
+
+  const e = await t.throwsAsync(() =>
+    upgrades.deployProxy(Greeter, ['hello'], { initialOwner: predeployedProxyAdminAddress }),
+  );
+  t.true(e.message.includes('`initialOwner` must not be a ProxyAdmin contract.'), e.message);
+  t.true(e.message.includes(predeployedProxyAdminAddress), e.message);
+});
+
+test('initial owner - must not be ownable', async t => {
+  const { Greeter, HasOwner } = t.context;
+
+  const defaultSigner = await ethers.provider.getSigner(0);
+  const hasOwner = await HasOwner.deploy(defaultSigner.address); // not actually a proxy admin, but it looks like one because it has an owner
+  const predeployedOwnableAddress = await hasOwner.getAddress();
+
+  const e = await t.throwsAsync(() =>
+    upgrades.deployProxy(Greeter, ['hello'], { initialOwner: predeployedOwnableAddress }),
+  );
+  t.true(e.message.includes('`initialOwner` must not be a ProxyAdmin contract.'), e.message);
+  t.true(e.message.includes(predeployedOwnableAddress), e.message);
+});
+
+test('initial owner - skip ProxyAdmin check', async t => {
+  const { Greeter, HasOwner } = t.context;
+
+  const defaultSigner = await ethers.provider.getSigner(0);
+  const hasOwner = await HasOwner.deploy(defaultSigner.address);
+  const predeployedOwnableAddress = await hasOwner.getAddress();
+
+  await upgrades.deployProxy(Greeter, ['hello'], {
+    initialOwner: predeployedOwnableAddress,
+    unsafeSkipProxyAdminCheck: true,
+  });
 });