+++ packages/plugin-hardhat/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 3.4.0 (2024-09-23)
+
+### Potentially breaking changes
+- Adds a check to ensure `initialOwner` is not a ProxyAdmin contract when deploying a transparent proxy. ([#1083](https://github.com/OpenZeppelin/openzeppelin-upgrades/pull/1083))
+
 ## 3.3.0 (2024-09-16)
 
 - Defender: Add `metadata` option. ([#1079](https://github.com/OpenZeppelin/openzeppelin-upgrades/pull/1079))